% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-involvement-survey.R
\docType{data}
\name{involvement_survey_str2s}
\alias{involvement_survey_str2s}
\title{Parent involvement survey: stratified, two-stage sample}
\format{
A data frame with 5,000 rows and 18 variables
}
\usage{
involvement_survey_str2s
}
\description{
An example dataset describing a stratified, multistage sample of 1,000 parents
of students with disabilities, from a population of 20,000.
The parent involvement survey measures a single key outcome:
whether "parents perceive that schools facilitate parent involvement
as a means of improving services and results for children with disabilities." \cr \cr
The sample was selected by sampling 5 schools from each of 20 districts,
and then sampling parents of 10 children in each sampled school.
The variable \code{BASE_WEIGHT} provides the base sampling weight.
The variable \code{SCHOOL_DISTRICT} was used for stratification,
and the variables \code{SCHOOL_ID} and \code{UNIQUE_ID} uniquely identify
the first and second stage sampling units (schools and parents).
The variables \code{N_SCHOOLS_IN_DISTRICT} and \code{N_STUDENTS_IN_SCHOOL}
can be used to provide finite population corrections.
}
\section{Fields}{


\describe{
\item{UNIQUE_ID}{A unique identifier for students}
\item{RESPONSE_STATUS}{Survey response/eligibility status: 'Respondent', 'Nonrespondent', 'Ineligble', 'Unknown'}
\item{WHETHER_PARENT_AGREES}{Parent agreement ('AGREE' or 'DISAGREE') for whether they perceive that schools facilitate parent involvement}
\item{SCHOOL_DISTRICT}{A unique identifier for school districts}
\item{SCHOOL_ID}{A unique identifier for schools, nested within districts}
\item{STUDENT_GRADE}{Student's grade level: 'PK', 'K', 1-12}
\item{STUDENT_AGE}{Student's age, measured in years}
\item{STUDENT_DISABILITY_CODE}{Code for student's disability category (e.g. 'VI' for 'Visual Impairments')}
\item{STUDENT_DISABILITY_CATEGORY}{Student's disability category (e.g. 'Visual Impairments')}
\item{STUDENT_SEX}{'Female' or 'Male'}
\item{STUDENT_RACE}{Seven-level code with descriptive label (e.g. 'AS7 (Asian)')}
\item{PARENT_HAS_EMAIL}{Whether parent has an e-mail address ('Has Email' vs 'No Email')}
\item{PARENT_HAS_EMAIL_BENCHMARK}{Population benchmark for category of \code{PARENT_HAS_EMAIL}}
\item{STUDENT_RACE_BENCHMARK}{Population benchmark for category of \code{STUDENT_RACE}}
\item{N_SCHOOLS_IN_DISTRICT}{Total number of schools in each district}
\item{N_STUDENTS_IN_SCHOOL}{Total number of students in each school}
\item{BASE_WEIGHT}{Sampling weight to use for weighted estimates}
\item{CONTACT_ATTEMPTS}{The number of contact attempts made for each case (ranges between 1 and 6)}
}
}

\examples{
# Load the data
involvement_survey_str2s

# Prepare the data for analysis with the 'survey' package

  library(survey)

  involvement_survey <- svydesign(
    data = involvement_survey_str2s,
    weights = ~ BASE_WEIGHT,
    strata =  ~ SCHOOL_DISTRICT,
    ids =     ~ SCHOOL_ID             + UNIQUE_ID,
    fpc =     ~ N_SCHOOLS_IN_DISTRICT + N_STUDENTS_IN_SCHOOL
  )

}
\keyword{datasets}
